Refactoring: add root-reducer
