+++ project/src/components/private-route/private-route.tsx
@@ -5,11 +5,11 @@ import {useSelector} from 'react-redux';
 import {State} from '../../types/state';
 
 function PrivateRoute({...props}: RouteProps): JSX.Element {
-  const authStatus = useSelector((state: State) => state.auth.status);
+  const isAuthorized = useSelector((state: State) => state.user.auth.status === AuthStatus.Auth);
 
   return (
     <Route {...props}>
-      {authStatus === AuthStatus.Auth ? props.children : <Redirect to={AppRoute.LoginScreen} />}
+      {isAuthorized ? props.children : <Redirect to={AppRoute.LoginScreen} />}
     </Route>
   );
 }
