+++ project/src/index.tsx
@@ -1,16 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {applyMiddleware, createStore} from '@reduxjs/toolkit';
-import thunk from 'redux-thunk';
-import {createAPI} from './services/api';
-import {reducer} from './store/reducer';
 import {Provider} from 'react-redux';
-import {composeWithDevTools} from 'redux-devtools-extension';
+import {configureStore} from '@reduxjs/toolkit';
+import {createAPI} from './services/api';
+import {rootReducer} from './store/root-reducer';
+import {requireAuthorization} from './store/user/user-actions';
+import {checkAuthAction} from './store/user/user-api-actions';
 import App from './components/app/app';
 import {AuthStatus, ToastParams} from './common/const';
-import {requireAuthorization} from './store/actions';
-import {ThunkAppDispatch} from './types/actions';
-import {checkAuthAction} from './store/api-actions';
 import {ToastContainer} from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 
@@ -18,12 +15,17 @@ const api = createAPI(() => {
   store.dispatch(requireAuthorization(AuthStatus.NoAuth));
 });
 
-const store = createStore(
-  reducer,
-  composeWithDevTools(applyMiddleware(thunk.withExtraArgument(api))),
-);
+const store = configureStore({
+  reducer: rootReducer,
+  middleware: (getDefaultMiddleware) =>
+    getDefaultMiddleware({
+      thunk: {
+        extraArgument: api,
+      },
+    }),
+});
 
-(store.dispatch as ThunkAppDispatch)(checkAuthAction());
+store.dispatch(checkAuthAction());
 
 ReactDOM.render(
   <React.StrictMode>
